Add Update-NinjaOneOrganisationDocument

+++ NinjaOne.build.ps1
@@ -1,55 +1,62 @@
 <#
     .SYNOPSIS
-        Build script for NinjaOne module - uses 'Invoke-Build' https://github.com/nightroman/Invoke-Build
+        Build script for NinjaOne module.
 #>
 Param (
-    [String]$Configuration = 'Development'
+    [String]$Configuration = 'Development',
+    [Switch]$UpdateHelp,
+    [Switch]$CopyModuleFiles,
+    [Switch]$Test,
+    [Switch]$UpdateManifest,
+    [Switch]$PublishModule,
+    [Switch]$Clean
 )
 
-#region Config: Use string mode when building.
-Set-StrictMode -Version Latest
-#endregion
+$ModuleName = 'NinjaOne'
 
-#region Task: Update the PowerShell Module Help Files.
+# Use strict mode when building.
+Set-StrictMode -Version Latest
+
+# Update the PowerShell Module Help Files.
 # Pre-requisites: PowerShell Module PlatyPS.
-task UpdateHelp {
-    Import-Module -Name 'NinjaOne' -Force
+
+if ($UpdateHelp) {
+    Import-Module -Name $ModuleName -Force
     Update-MarkdownHelp -Path "$($PSScriptRoot)\Docs\Markdown"
     New-ExternalHelp -Path "$($PSScriptRoot)\Docs\Markdown" -OutputPath "$($PSScriptRoot)\Docs\en_GB" -Force
 }
-#endregion
 
-#region Task: Copy PowerShell Module files to output folder for release on PSGallery
-task CopyModuleFiles {
+# Copy PowerShell Module files to output folder for release on PSGallery
+if ($CopyModuleFiles) {
     # Copy Module Files to Output Folder
-    if (-not (Test-Path "$($PSScriptRoot)\Output\NinjaOne")) {
-        New-Item -Path "$($PSScriptRoot)\Output\NinjaOne" -ItemType Directory | Out-Null
+    if (-not (Test-Path "$($PSScriptRoot)\Output\$ModuleName")) {
+        New-Item -Path "$($PSScriptRoot)\Output\$ModuleName" -ItemType Directory | Out-Null
     }
-    Copy-Item -Path "$($PSScriptRoot)\Private\" -Filter *.* -Recurse -Destination "$($PSScriptRoot)\Output\NinjaOne" -Force
-    Copy-Item -Path "$($PSScriptRoot)\Public\" -Filter *.* -Recurse -Destination "$($PSScriptRoot)\Output\NinjaOne" -Force
+    Copy-Item -Path "$($PSScriptRoot)\Classes\" -Filter *.* -Recurse -Destination "$($PSScriptRoot)\Output\$ModuleName" -Force
+    Copy-Item -Path "$($PSScriptRoot)\Data\" -Filter *.* -Recurse -Destination "$($PSScriptRoot)\Output\$ModuleName" -Force
+    Copy-Item -Path "$($PSScriptRoot)\Private\" -Filter *.* -Recurse -Destination "$($PSScriptRoot)\Output\$ModuleName" -Force
+    Copy-Item -Path "$($PSScriptRoot)\Public\" -Filter *.* -Recurse -Destination "$($PSScriptRoot)\Output\$ModuleName" -Force
 
-    #Copy Module Manifest files
+    # Copy module, manifest and scaffold files
     Copy-Item -Path @(
         "$($PSScriptRoot)\LICENSE.md"
         "$($PSScriptRoot)\CHANGELOG.md"
         "$($PSScriptRoot)\README.md"
-        "$($PSScriptRoot)\NinjaOne.psd1"
-        "$($PSScriptRoot)\NinjaOne.psm1"
-    ) -Destination "$($PSScriptRoot)\Output\NinjaOne" -Force
+        "$($PSScriptRoot)\$ModuleName.psd1"
+        "$($PSScriptRoot)\$ModuleName.psm1"
+    ) -Destination "$($PSScriptRoot)\Output\$ModuleName" -Force
 }
-#endregion
 
-#region Task: Run all Pester tests in folder .\Tests
-task Test {
+# Run all Pester tests in folder .\Tests
+if ($Test) {
     $Result = Invoke-Pester "$($PSScriptRoot)\Tests" -PassThru
     if ($Result.FailedCount -gt 0) {
         throw 'Pester tests failed'
     }
 }
-#endregion
 
-#region Task: Update the Module Manifest file with info from the Changelog.
-task UpdateManifest {
+# Update the Module Manifest file with info from the Changelog.
+if ($UpdateManifest) {
     # Import PlatyPS. Needed for parsing the versions in the Changelog.
     Import-Module -Name PlatyPS
 
@@ -58,52 +65,53 @@ task UpdateManifest {
     $MarkdownObject = [Markdown.MAML.Parser.MarkdownParser]::new()
     [regex]$Regex = '\d\.\d\.\d'
     $Versions = $Regex.Matches($MarkdownObject.ParseString($CHANGELOG).Children.Spans.Text) | ForEach-Object { $_.Value }
-    ($Versions | Measure-Object -Maximum).Maximum
+    $ChangeLogVersion = ($Versions | Measure-Object -Maximum).Maximum
 
-    $ManifestPath = "$($PSScriptRoot)\NinjaOne.psd1"
+    $ManifestPath = "$($PSScriptRoot)\$ModuleName.psd1"
 
     # Start by importing the manifest to determine the version, then add 1 to the Build
     $Manifest = Test-ModuleManifest -Path $ManifestPath
     [System.Version]$Version = $Manifest.Version
-    [String]$NewVersion = New-Object -TypeName System.Version -ArgumentList ($Version.Major, $Version.Minor, ($Version.Build + 1))
-    Write-Output -InputObject ("New Module version: $($NewVersion)")
+
+    if ($ChangeLogVersion -eq $Version) {
+        Throw 'No new version found in CHANGELOG.md'
+    }
+
+    Write-Output -InputObject ("Current Module Version: $($Version)")
+    Write-Output -InputObject ("New Module version: $($ChangeLogVersion)")
 
     # Update Manifest file with Release Notes
     $CHANGELOG = Get-Content -Path "$($PSScriptRoot)\CHANGELOG.md"
     $MarkdownObject = [Markdown.MAML.Parser.MarkdownParser]::new()
     $ReleaseNotes = ((($MarkdownObject.ParseString($CHANGELOG).Children.Spans.Text) -Match '\d\.\d\.\d') -Split ' - ')[1]
 
-    #Update Module with new version
-    Update-ModuleManifest -ModuleVersion $NewVersion -Path "$($PSScriptRoot)\NinjaOne.psd1" -ReleaseNotes $ReleaseNotes
+    # Update Module with new version
+    Update-ModuleManifest -ModuleVersion $ChangeLogVersion -Path "$($PSScriptRoot)\$ModuleName.psd1" -ReleaseNotes $ReleaseNotes
 }
-#endregion
 
-#region Task: Publish Module to PowerShell Gallery
-task PublishModule -if ($Configuration -eq 'Production') {
+# Publish Module to PowerShell Gallery
+if ($PublishModule -and $Configuration -eq 'Production') {
     Try {
         # Build a splat containing the required details and make sure to Stop for errors which will trigger the catch
         $params = @{
-            Path = ("$($PSScriptRoot)\Output\NinjaOne")
-            NuGetApiKey = (Get-AzKeyVaultSecret -VaultName $ENV:PSGalleryVault -Name $ENV:PSGallerySecret -AsPlainText)
+            Path = ("$($PSScriptRoot)\Output\$ModuleName")
+            NuGetApiKey = $ENV:TF_BUILD ? $ENV:PSGalleryAPIKey : (Get-AzKeyVaultSecret -VaultName $ENV:PSGalleryVault -Name $ENV:PSGallerySecret -AsPlainText) # If running in Azure DevOps, use the Environment Variable, otherwise use the Key Vault
             ErrorAction = 'Stop'
         }
+        $ManifestPath = "$($PSScriptRoot)\$ModuleName.psd1"
+        $Manifest = Test-ModuleManifest -Path $ManifestPath
+        [System.Version]$Version = $Manifest.Version
         Publish-Module @params
-        Write-Output -InputObject ('NinjaOne PowerShell Module published to the PowerShell Gallery')
+        Write-Output -InputObject ("$ModuleName PowerShell Module version $($Version) published to the PowerShell Gallery")
     } Catch {
         Throw $_
     }
 }
-#endregion
 
-#region Task: Clean up Output folder
-task Clean {
+# Clean up Output folder
+if ($Clean) {
     # Clean output folder
     if ((Test-Path "$($PSScriptRoot)\Output")) {
         Remove-Item -Path "$($PSScriptRoot)\Output" -Recurse -Force
     }
 }
-#endregion
-
-#region Default Task. Runs Clean, Test, CopyModuleFiles Tasks
-task . Clean, Test, CopyModuleFiles, PublishModule, Clean
-#endregion
